fix: CI fix
