+++ packages/agents/chain-abstraction/src/libs/index.ts
@@ -1,3 +1,2 @@
 export * from "./origin";
 export * from "./shared";
-export * from "../helpers/swapquote";
